Moved `ecs` mod to its own file
